+++ src/modules/controls/orientation/components/Orientation.vue
@@ -72,6 +72,7 @@ export default {
     mounted () {
         this.addElement();
         this.checkWFS();
+
     },
     methods: {
         ...mapMutations("controls/orientation", Object.keys(mutations)),