+++ src/modules/tools/filterGeneral/store/indexFilterGeneral.js
@@ -1,0 +1,12 @@
+import state from "./stateFilterGeneral";
+import actions from "./actionsFilterGeneral";
+import getters from "./gettersFilterGeneral";
+import mutations from "./mutationsFilterGeneral";
+
+export default {
+    namespaced: true,
+    state: {...state},
+    mutations,
+    actions,
+    getters
+};